backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-allauth==0.57.0
dj-rest-auth==5.0.2
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-decouple==3.8
drf-yasg==1.21.7
psycopg2-binary==2.9.9
django-filter==23.5
django-extensions==3.2.3
Pillow==10.1.0